+++ packages/adapters/cache/src/lib/entities/channel.ts
@@ -1,0 +1,7 @@
+export type SubscriptionCallback = (msg: any, err?: any) => void;
+
+export type Subscriptions = Map<string, SubscriptionCallback>;
+
+export enum StoreChannel {
+  NewPreparedTx = "new-prepared-tx",
+}
