+++ packages/adapters/cache/src/lib/entities/index.ts
@@ -1,2 +1,3 @@
-export * from "./config";
 export * from "./cache";
+export * from "./channel";
+export * from "./store";
