+++ packages/adapters/cache/src/lib/entities/store.ts
@@ -2,5 +2,3 @@ import { Redis } from "ioredis";
 import { Logger } from "@connext/nxtp-utils";
 
 export type StoreManagerParams = { redis: { url: string; instance?: Redis }; logger: Logger; };
-
-export type CacheParams = { url: string; };
