+++ packages/contracts/typechain-types/index.ts
@@ -17,7 +17,6 @@ export type { IERC20Extended } from "./IERC20Extended";
 export type { IERC20Minimal } from "./IERC20Minimal";
 export type { IFulfillInterpreter } from "./IFulfillInterpreter";
 export type { IPriceOracle } from "./IPriceOracle";
-export type { IRouterFactory } from "./IRouterFactory";
 export type { IStableSwap } from "./IStableSwap";
 export type { IWrapped } from "./IWrapped";
 export type { FulfillInterpreter } from "./FulfillInterpreter";
@@ -54,7 +53,6 @@ export type { ITransactionManager } from "./ITransactionManager";
 export type { IWeth } from "./IWeth";
 export type { PriceOracle } from "./PriceOracle";
 export type { ProposedOwnable } from "./ProposedOwnable";
-export type { RouterFactory } from "./RouterFactory";
 export type { StableSwap } from "./StableSwap";
 export type { Counter } from "./Counter";
 export type { FeeERC20 } from "./FeeERC20";
@@ -84,7 +82,6 @@ export { IERC20Extended__factory } from "./factories/IERC20Extended__factory";
 export { IERC20Minimal__factory } from "./factories/IERC20Minimal__factory";
 export { IFulfillInterpreter__factory } from "./factories/IFulfillInterpreter__factory";
 export { IPriceOracle__factory } from "./factories/IPriceOracle__factory";
-export { IRouterFactory__factory } from "./factories/IRouterFactory__factory";
 export { IStableSwap__factory } from "./factories/IStableSwap__factory";
 export { IWrapped__factory } from "./factories/IWrapped__factory";
 export { FulfillInterpreter__factory } from "./factories/FulfillInterpreter__factory";
@@ -121,7 +118,6 @@ export { ITransactionManager__factory } from "./factories/ITransactionManager__f
 export { IWeth__factory } from "./factories/IWeth__factory";
 export { PriceOracle__factory } from "./factories/PriceOracle__factory";
 export { ProposedOwnable__factory } from "./factories/ProposedOwnable__factory";
-export { RouterFactory__factory } from "./factories/RouterFactory__factory";
 export { StableSwap__factory } from "./factories/StableSwap__factory";
 export { Counter__factory } from "./factories/Counter__factory";
 export { FeeERC20__factory } from "./factories/FeeERC20__factory";
